why this ass isn't work :(((

+++ hd/Integer.hpp
@@ -84,6 +84,6 @@ public:
 
 private:
 	bool sign_;
-	unsigned value_;
+	long value_;
 
 };
